requirements-web.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-chess>=1.11

